Tcl/requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.2
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
cryptography==3.2.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.44.0
django-braces==1.14.0
django-cognito-jwt==0.0.3
django-cors-headers==3.2.1
django-extensions==3.1.0
django-filter==2.2.0
django-graphql-auth==0.3.10
django-graphql-jwt==0.3.0
django-storages==1.10.1
djangorestframework==3.12.2
ecdsa==0.14.1
envs==1.3
future==0.18.2
gql==2.0.0
graphene==2.1.8
graphene-django==2.10.1
graphene-file-upload==1.2.2
graphene-sqlalchemy==2.3.0
graphene-sqlalchemy-filter==1.11.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.10
isort==4.3.21
Jinja2==2.11.2
jmespath==0.10.0
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mailchimp-marketing==3.0.25
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
ordereddict==1.1
pbr==5.4.5
phonenumbers==8.12.10
Pillow==8.0.0
promise==2.3
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyactiveresource==2.2.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycognito==0.1.4
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pylint-django==2.0.15
pylint-flask==0.6
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-decouple==3.3
python-jose==3.2.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-aws4auth==1.0.1
requests-oauthlib==1.3.0
rsa==4.6
Rx==1.6.1
s3transfer==0.3.3
ShopifyAPI==8.0.1
singledispatch==3.4.0.3
six==1.15.0
SQLAlchemy==1.3.17
sqlalchemy-migrate==0.13.0
SQLAlchemy-serializer==1.3.4.4
sqlparse==0.3.1
Tempita==0.5.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
WooCommerce==2.1.1
wrapt==1.12.1
